Modified script to avoid recreation of WWW folder inside iOS project folder.
On running iOS code, The WWW folder is recreated inside the iOS project folder,
that may lead to a confusion for the developers.
This script will first check whether the WWW folder is present inside the iOS project folder and if it's not there,
check outside the iOS folder(PhoneGap project folder).
If the WWW folder is absent in both cases it's created inside the iOS project folder.